refactor: move delivery module to novo/backend/modulos/delivery
- Move delivery module from service/backend to novo/backend/modulos
- Update all imports to use modulos.delivery.* paths
- Reorganize BDD tests to novo/bdd/features structure
- Keep delivery architecture and maintain test coverage
- Align with novo monorepo module structure
