+++ novo/backend/modulos/delivery/domain/ports.py
