+++ novo/backend/modulos/delivery/infrastructure/repositories.py
@@ -1,10 +1,10 @@
 from uuid import UUID
 from typing import Optional
 
-from delivery.domain.entities import Deliverer, Order
-from delivery.domain.enums import DelivererStatus, OrderStatus
-from delivery.domain.ports import DelivererRepository, OrderRepository
-from delivery.infrastructure.models.deliverers_model import DelivererModel, OrderModel
+from modulos.delivery.domain.entities import Deliverer, Order
+from modulos.delivery.domain.enums import DelivererStatus, OrderStatus
+from modulos.delivery.domain.ports import DelivererRepository, OrderRepository
+from modulos.delivery.infrastructure.models.deliverers_model import DelivererModel, OrderModel
 
 
 class DelivererRepositoryImpl(DelivererRepository):
